+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 + Check if public IP is different from DNS record
 + Add browse zones test
 + Fix code smells
++ Add browse dns records test
 
 ## 0.2.2
 + Fix yaml